ChatVideoPose/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
transformers>=4.30.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
tqdm>=4.65.0
scipy>=1.10.0
pillow>=9.5.0
mediapipe>=0.10.0  # For basic pose estimation
huggingface-hub>=0.16.0